refactor(lynx-react): move get-seed-class-name to utils
Relocate get-seed-class-name.ts into the utils domain folder so
top-level src/ only contains the package entry point.
